feat: expand authentication package and migrate components from customer app
- Define `AuthEndpoints` and `AuthStorageKeys` in the authentication package.
- Migrate `AuthErrorType` and `ForgotPasswordRequest` from `apps/customer` to `packages/authentication`.
- Enhance `AuthErrorType` with a `toLocalizedString` method for localized error messages.
- Add `ForgotPasswordResetRequest` data model.
- Update the `Store` entity to reference the `Address` entity from the authentication package.
