+++ packages/authentication/lib/core/constants/auth_constants.dart
